Strapi Server
Here's the server, use "yarn develop" or "npm develop" or whatever command to run it. You have to extract in someplace convient, I just did it on my desktop in it's own folder. Then open the folder in CMD and type "yarn develop" if you have yarn installed or "npm or npx develop" if you have npm (node.js).